+++ hadamard.cpp
@@ -8,7 +8,7 @@
  * @param z puntero a vector resultado del producto elemento a elemento de x y y.
  * 
  */
-void hadamard(float *x, float *y, int tm, float *z)
+void hadamard(double *x, double *y, int tm, double *z)
 {
 	for(int i=0;i<tm;i++)
 		z[i]=x[i]*y[i];
